Fix loading Timecop when Date is not loaded, but Time is
When `require "time"` is used to load Time, but no `require "date"` is
issued, Timecop fails to load with `require "timecop"`. This is because
`require "time"` loads "date/format", which ultimately misleads Timecop
to think that Date is fully loaded and thus causing Timecop to make
changes to Date and DateFormat classes. This patch fixes the problem by
making Timecop not to change Date and DateFormat classes if the classes
do not respond to the methods that Timecop would modify.