Documentation tweak for syntax of RSpec usage
The around syntax is better than using separate before and after syntax when
integrating with RSpec and the docs now reflect that. It is more terse and
comes with the benefit of being executed in a block with an ensure (meaning)
time won't accidentally leak between specs and be frozen in the past or future.