Fix testing timezone dates
The test_timezones_apply_dates test actually fails to test
timezones. It is comparing the local time with the time zone
specific time. Therefore the test actually only "passes" if the
time zone is to the "west" of the local time zone. A person in
the America/Los_Angeles time zone will receive a failure when
running the current test as the Central Time zone is to the "east".

Change the test to use Time.zone.now for the comparison date