Calculate travel_offset to align with the precision of argument to Timecop.travel (#421)
* Calculate travel_offset to align with clock precision

Subtracting two `Time`s returns a `Float`, which may not be accurate
down to subsecond resolution. Because `Float`s are stored as double-
precision values (IEEE 754), they can have resolutions much higher than
the typical minimum clock precision of 10e-9 seconds. Which can result
in two `Time` object not comparing as equal when they are the same down
to the nanosecond, when one has had a travel_offset applied to it.

* Use the realtime clock's resolution when testing travel_offset's alignment

* Improve failure message for travel_offset alignment test

* Test that time_offset keeps precision of very high times passed to Timecop.travel

* Mention this change in History.md